staging: refactor docs for clearer scoping
